Moved most tree specific code into class
