+++ src/prototree/train.py
@@ -8,7 +8,6 @@ from torch.utils.data import DataLoader
 from tqdm import tqdm
 
 from prototree.models import ProtoTree
-from prototree.node import Leaf, Node, NodeProbabilities
 
 log = logging.getLogger(__name__)
 
@@ -40,11 +39,11 @@ def train_epoch(
         optimizer.step()
 
         smoothing_factor = 1 - 1 / n_batches
-        model.eval()
-        update_leaf_distributions(
-            model.tree_section.tree_root, y, logits.detach(), node_to_prob, smoothing_factor
+        model.tree_section.update_leaf_distributions(
+            y, logits.detach(), node_to_prob, smoothing_factor
         )
 
+        model.eval()
         y_pred = torch.argmax(logits, dim=1)
         acc = torch.sum(y_pred == y).item() / len(x)
         tqdm_loader.set_postfix_str(f"batch: loss={loss.item():.5f}, {acc=:.5f}")
@@ -64,65 +63,3 @@ def train_epoch(
         "loss": avg_loss,
         "train_accuracy": avg_acc,
     }
-
-
-@torch.no_grad()
-def update_leaf_distributions(
-    root: Node,
-    y_true: torch.Tensor,
-    logits: torch.Tensor,
-    node_to_prob: dict[Node, NodeProbabilities],
-    smoothing_factor: float,
-):
-    """
-    :param root:
-    :param y_true: shape (batch_size)
-    :param logits: shape (batch_size, num_classes)
-    :param node_to_prob:
-    :param smoothing_factor:
-    """
-    batch_size, num_classes = logits.shape
-
-    y_true_one_hot = F.one_hot(y_true, num_classes=num_classes)
-    y_true_logits = torch.log(y_true_one_hot)
-
-    for leaf in root.leaves:
-        update_leaf(leaf, node_to_prob, logits, y_true_logits, smoothing_factor)
-
-
-def update_leaf(
-    leaf: Leaf,
-    node_to_prob: dict[Node, NodeProbabilities],
-    logits: torch.Tensor,
-    y_true_logits: torch.Tensor,
-    smoothing_factor: float,
-):
-    """
-    :param leaf:
-    :param node_to_prob:
-    :param logits: of shape (batch_size, num_classes)
-    :param y_true_logits: of shape (batch_size, num_classes)
-    :param smoothing_factor:
-    :return:
-    """
-    # shape (batch_size, 1)
-    log_p_arrival = node_to_prob[leaf].log_p_arrival.unsqueeze(1)
-    # shape (num_classes). Not the same as logits, which has (batch_size, num_classes)
-    leaf_logits = leaf.y_logits()
-
-    # TODO: y_true_logits is mostly -Inf terms (the rest being 0s) that won't contribute to the total, and we are also
-    #  summing together tensors of different shapes. We should be able to express this more clearly and efficiently by
-    #  taking advantage of this sparsity.
-    log_dist_update = torch.logsumexp(
-        log_p_arrival + leaf_logits + y_true_logits - logits,
-        dim=0,
-    )
-
-    dist_update = torch.exp(log_dist_update)
-
-    # This scaling (subtraction of `-1/n_batches * c` in the ProtoTree paper) seems to be a form of exponentially
-    # weighted moving average, designed to ensure stability of the leaf class probability distributions (
-    # leaf.dist_params), by filtering out noise from minibatching in the optimization.
-    leaf.dist_params.mul_(smoothing_factor)
-
-    leaf.dist_params.add_(dist_update)
